moved queries to a query factory, implemented v2 queries for berkeley compatibility
