+++ sample.env
@@ -15,6 +15,8 @@ BLOCK_DB_QUERY_HOST=db
 BLOCK_DB_QUERY_PORT=5432
 BLOCK_DB_QUERY_NAME=archive
 BLOCK_DB_QUERY_REQUIRE_SSL=true
+#database version: v1 = mainnet; v2 = berkeley
+BLOCK_DB_VERSION="v1" 
 
 #Staking Ledger QUERY Database settings
 LEDGER_DB_QUERY_USER=postgres
